requirements.txt
anyio==4.7.0
async-timeout==5.0.1
bcrypt==4.2.1
click==8.1.8
Deprecated==1.2.15
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.95.0
fastapi-limiter==0.1.6
greenlet==3.1.1
h11==0.14.0
idna==3.10
limits==3.14.1
packaging==24.2
passlib==1.7.4
psycopg2-binary==2.9.6
pyasn1==0.6.1
pydantic==1.10.10
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.20
redis==4.6.0
rsa==4.9
setuptools==75.6.0
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
SQLAlchemy==2.0.14
starlette==0.26.1
typing_extensions==4.12.2
uvicorn==0.22.0
wrapt==1.17.0
